+++ src/__tests__/__fixtures__/instanceWithDependentSteps/src/getStepStartStates.ts
@@ -1,0 +1,13 @@
+export default function getStepStartStates() {
+  return {
+    'fetch-accounts': {
+      disabled: false,
+    },
+    'fetch-users': {
+      disabled: false,
+    },
+    'fetch-groups': {
+      disabled: false,
+    },
+  };
+}